agent/pyproject.toml
[project]
name = "jatango-voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[openai,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "httpx",
]
